+++ apm-sniffer/apm-sdk-plugin/cxf-3.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/cxf/CxfInstrumentation.java
@@ -16,7 +16,7 @@
  *
  */
 
-package cxf;
+package org.apache.skywalking.apm.plugin.cxf;
 
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.matcher.ElementMatcher;
@@ -28,11 +28,10 @@ import org.apache.skywalking.apm.agent.core.plugin.match.NameMatch;
 
 import static net.bytebuddy.matcher.ElementMatchers.named;
 
-public class CxfDynamicInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
+public class CxfInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
 
-
-    private static final String ENHANCE_CLASS = "org.apache.cxf.transport.http.URLConnectionHTTPConduit";
-    private static final String INTERCEPT_CLASS = "org.apache.skywalking.apm.plugin.cxf.CxfDynamicInterceptor";
+    private static final String ENHANCE_CLASS = "org.apache.axis.transport.http.HTTPSender";
+    private static final String INTERCEPT_CLASS = "org.apache.skywalking.apm.plugin.CxfInterceptor";
 
     @Override
     protected ClassMatch enhanceClass() {
@@ -47,11 +46,11 @@ public class CxfDynamicInstrumentation extends ClassInstanceMethodsEnhancePlugin
     @Override
     protected InstanceMethodsInterceptPoint[] getInstanceMethodsInterceptPoints() {
         return new InstanceMethodsInterceptPoint[]{
-                //直接调用wsdl
+                //通过wsdl生成客户端进行调用
                 new InstanceMethodsInterceptPoint() {
                     @Override
                     public ElementMatcher<MethodDescription> getMethodsMatcher() {
-                        ElementMatcher<MethodDescription> em = named("createConnection");
+                        ElementMatcher<MethodDescription> em = named("invoke");
                         return em;
                     }
